Task admin modal windows

+++ MyBook/Controllers/AdminProfileController.cs
@@ -10,7 +10,7 @@ using System.Security.Claims;
 using Type = MyBook.Entities.Type;
 namespace MyBook.Controllers
 {
-    /*[Authorize(Policy = "AdminsOnly")]*/
+    [Authorize(Policy = "AdminsOnly")]
     public class AdminProfileController : Controller
     {
         private readonly IGenericRepository<Type> _typeRepository;
@@ -71,12 +71,18 @@ namespace MyBook.Controllers
                 CurrentPage = page
             });
         }
+
         #region SubscrType
-        public async Task<IActionResult> DeleteSubscrType(Type type)
+        public async Task<IActionResult> DeleteSubscription(int id)
         {
+            var modalModel = new ModalsViewModel { ControllerName = "AdminProfile", ActionName = "ShowCurrent", CurrentPage = "Subscription" };
+            var type = await _typeRepository.FindById(id);
+            if (type == null)
+            {
+                return RedirectToAction("Error", "Modals",  modalModel );
+            }
             await _typeRepository.Remove(type);
-            var page = "Subscription";
-            return RedirectToAction("ShowCurrent", new { page });
+            return RedirectToAction("Successful", "Modals", modalModel );
         }
 
         public IActionResult EditSubscriptionModal(EditSubscrViewModel model)
@@ -95,8 +101,8 @@ namespace MyBook.Controllers
             {
                 var type = new Type { Description = model.Description, Price = model.Price, TypeName = model.TypeName };
                 await _typeRepository.Create(type);
-                var page = "Subscription";
-                return RedirectToAction("ShowCurrent", new { page });
+                var modalModel = new ModalsViewModel { ControllerName = "AdminProfile", ActionName = "ShowCurrent", CurrentPage = "Subscription" };
+                return RedirectToAction("Successful", "Modals",  modalModel );
             }
             return View("AddSubscriptionModal", model);
         }
@@ -115,8 +121,8 @@ namespace MyBook.Controllers
                 type.Price = model.Price;
                 type.TypeName = model.TypeName;
                 await _typeRepository.Update(type, null);
-                var page = "Subscription";
-                return RedirectToAction("ShowCurrent", new { page });
+                var modalModel = new ModalsViewModel { ControllerName = "AdminProfile", ActionName = "ShowCurrent", CurrentPage = "Subscription" };
+                return RedirectToAction("Successful", "Modals",  modalModel);
             }
             return View("EditSubscriptionModal", model);
         }
@@ -124,12 +130,14 @@ namespace MyBook.Controllers
         #endregion
 
         #region Author
-        //TODO search author
-        public async Task<IActionResult> DeleteAuthor(Author author)
+        public async Task<IActionResult> DeleteAuthor(int id)
         {
+            var modalModel = new ModalsViewModel { ControllerName = "AdminProfile", ActionName = "ShowCurrent", CurrentPage = "Author" };
+            var author = await _authorRepository.FindById(id);
+            if (author == null)
+                return RedirectToAction("Error", "Modals", modalModel );
             await _authorRepository.Remove(author);
-            var page = "Author";
-            return RedirectToAction("ShowCurrent", new { page });
+            return RedirectToAction("Successful", "Modals", modalModel );
         }
         public IActionResult EditAuthorModal(EditAuthorViewModel model)
         {
@@ -153,8 +161,8 @@ namespace MyBook.Controllers
                     await _genericRepository.CreateAll(new List<object>() { author, imageLink });
                 }
                 else await _authorRepository.Create(author);
-                var page = "Author";
-                return RedirectToAction("ShowCurrent", new { page });
+                var modalModel = new ModalsViewModel { ControllerName = "AdminProfile", ActionName = "ShowCurrent", CurrentPage = "Author" };
+                return RedirectToAction("Successful", "Modals", modalModel );
             }
             return View("AddAuthorModal", model);
         }
@@ -179,8 +187,8 @@ namespace MyBook.Controllers
                     author.ImgLinks = new List<ImgLink> { imageLink };
                 }
                 await _authorRepository.Update(author, null);
-                var page = "Author";
-                return RedirectToAction("ShowCurrent", new { page });
+                var modalModel = new ModalsViewModel { ControllerName = "AdminProfile", ActionName = "ShowCurrent", CurrentPage = "Author" };
+                return RedirectToAction("Successful", "Modals", modalModel );
             }
             return View("EditAuthorModal", model);
         }
@@ -188,13 +196,14 @@ namespace MyBook.Controllers
         #endregion
 
         #region Book
-        //TODO search book
-        public async Task<IActionResult> DeleteBook(Book book)
+        public async Task<IActionResult> DeleteBook(int id)
         {
-            //TODO fix with other links
+            var modalModel = new ModalsViewModel { ControllerName = "AdminProfile", ActionName = "ShowCurrent", CurrentPage = "Book" };
+            var book = await _bookRepository.FindById(id);
+            if (book == null)
+                return RedirectToAction("Error", "Modals", modalModel);
             await _bookRepository.Remove(book);
-            var page = "Book";
-            return RedirectToAction("Index", new { page });
+            return RedirectToAction("Successful", "Modals", modalModel);
         }
 
         public IActionResult EditBookModal(EditBookViewModel model)
@@ -247,12 +256,12 @@ namespace MyBook.Controllers
                 entities.Add(book);
                 entities.Add(desc);
                 await _genericRepository.CreateAll(entities);
-                var page = "Book";
-                return RedirectToAction("ShowCurrent", new { page });
+                var modalModel = new ModalsViewModel { ControllerName = "AdminProfile", ActionName = "ShowCurrent", CurrentPage = "Book" };
+                return RedirectToAction("Successful", "Modals", modalModel);
             }
             return View("AddBookModal", model);
         }
-        //TODO почему то дублируются авторы и жанры
+
         public async Task<IActionResult> EditBook(EditBookViewModel model)
         {
             if (ModelState.IsValid && model.BookId != null)
@@ -300,8 +309,8 @@ namespace MyBook.Controllers
                 entities.Add(desc);
                 await _genericRepository.CreateAll(entities);
                 await _bookRepository.Update(book);
-                var page = "Book";
-                return RedirectToAction("ShowCurrent", new { page });
+                var modalModel = new ModalsViewModel { ControllerName = "AdminProfile", ActionName = "ShowCurrent", CurrentPage = "Book" };
+                return RedirectToAction("Successful", "Modals", modalModel);
             }
             return View("EditBookModal", model);
         }
@@ -320,11 +329,14 @@ namespace MyBook.Controllers
             return View(new EditCenterViewModel());
         }
 
-        public async Task<IActionResult> DeleteBookCenter(BookCenter bookCenter)
+        public async Task<IActionResult> DeleteBookCenter(int id)
         {
+            var modalModel = new ModalsViewModel { ControllerName = "AdminProfile", ActionName = "ShowCurrent", CurrentPage = "BookCenter" };
+            var bookCenter = await _bookCenterRepository.FindById(id);
+            if (bookCenter == null)
+                return RedirectToAction("Error", "Modals", modalModel);
             await _bookCenterRepository.Remove(bookCenter);
-            var page = "BookCenter";
-            return RedirectToAction("ShowCurrent", new { page });
+            return RedirectToAction("Successful", "Modals", modalModel);
         }
 
         public async Task<IActionResult> AddBookCenter(EditCenterViewModel model)
@@ -333,8 +345,8 @@ namespace MyBook.Controllers
             {
                 var center = new BookCenter { Address = model.Address, Name = model.Name, Description = model.Description, Phone = model.Phone };
                 await _bookCenterRepository.Create(center);
-                var page = "BookCenter";
-                return RedirectToAction("ShowCurrent", new { page });
+                var modalModel = new ModalsViewModel { ControllerName = "AdminProfile", ActionName = "ShowCurrent", CurrentPage = "BookCenter" };
+                return RedirectToAction("Successful", "Modals", modalModel);
             }
             return View("AddBookCenterModal", model);
         }
@@ -354,8 +366,8 @@ namespace MyBook.Controllers
                 center.Description = model.Description;
                 center.Phone = model.Phone;
                 await _bookCenterRepository.Update(center, null);
-                var page = "BookCenter";
-                return RedirectToAction("ShowCurrent", new { page });
+                var modalModel = new ModalsViewModel { ControllerName = "AdminProfile", ActionName = "ShowCurrent", CurrentPage = "BookCenter" };
+                return RedirectToAction("Successful", "Modals", modalModel);
             }
             return View("EditBookCenterModal", model);
         }
@@ -363,14 +375,14 @@ namespace MyBook.Controllers
         #endregion
 
         #region User
-        //TODO search User
         public async Task<IActionResult> DeleteUserSubscr(string id, int subscrId)
         {
+            var modalModel = new ModalsViewModel { ControllerName = "AdminProfile", ActionName = "ShowCurrent", CurrentPage = "User" };
             var userSubscr = _userSubscrRepository.GetUserSubscr(id, subscrId);
-            if (userSubscr == null) { } //TODO show error
+            if (userSubscr == null)
+                return RedirectToAction("Error", "Modals", modalModel);
             await _userSubscrRepository.Remove(userSubscr);
-            var page = "User";
-            return RedirectToAction("ShowCurrent", new { page });
+            return RedirectToAction("Successful", "Modals", modalModel);
         }
         public IActionResult AddSubscrModal(string userId)
         {
@@ -381,8 +393,8 @@ namespace MyBook.Controllers
         {
             var user = await _userManager.FindByIdAsync(userId);
             await _userManager.AddClaimAsync(user, new Claim(ClaimTypes.Role, "Admin"));
-            var page = "User";
-            return RedirectToAction("ShowCurrent", new { page });
+            var modalModel = new ModalsViewModel { ControllerName = "AdminProfile", ActionName = "ShowCurrent", CurrentPage = "User" };
+            return RedirectToAction("Successful", "Modals", modalModel);
         }
 
 
@@ -433,8 +445,8 @@ namespace MyBook.Controllers
                 var userSubscr = new UserSubscr { Subscription = subscr, UserId = model.UserID };
                 subscr.UserSubscr = userSubscr;
                 await _genericRepository.CreateAll(new List<object>() { subscr, userSubscr, });
-                var page = "User";
-                return RedirectToAction("ShowCurrent", new { page });
+                var modalModel = new ModalsViewModel { ControllerName = "AdminProfile", ActionName = "ShowCurrent", CurrentPage = "User" };
+                return RedirectToAction("Successful", "Modals", modalModel);
             }
             return View("AddSubscriptionToUserModal", model);
         }
@@ -466,7 +478,7 @@ namespace MyBook.Controllers
         private async Task<List<User>> GetUsers()
         {
             var admins = await _userManager.GetUsersForClaimAsync(new Claim(ClaimTypes.Role, "Admin"));
-            return _userRepository.GetUsersWithSubscr().Where(user=>!admins.Contains(user)).ToList();
+            return _userRepository.GetUsersWithSubscr().Where(user => !admins.Contains(user)).ToList();
         }
 
         private List<BookCenter> GetCenters()
